Linux compilation issues fix

+++ include/TmxTile.h
@@ -74,7 +74,7 @@ namespace Tmx
         /// Returns the number of frames of the animation. If the tile is not animated, returns 0.
         int GetFrameCount() const
         {
-            return (int)frames.size();
+            return frames.size();
         }
 
         /// Returns the total duration of the animation, in milliseconds,
